Update m_pendingFloatBoxNumsBeforeCurrentLine if it's completely new line
If every time new line box is generated m_pendingFloatBoxNumsBeforeCurrentLine is updated, the problem can happen when there're pending inline
boxes. Because the intend of m_pendingFloatBoxNumsBeforeCurrentLine is to separate the float boxes inside the same vector which all the
pending float boxes resides in, for the original intend, the condition `m_pendingInlineBoxes.size() == 0` was added.
Also, the last paremeter in the function `insertPendingFloatBoxes` was removed becuase it is redundant.