Implement MarginCollapse dynamically
According to whether `clear` is affected, margin of block can or cannot participate in margin-collapse.
The fact that `clear` is affected only can be known after trying layout actually. Therefore we should layout first assuming estimated
margin-collapse result and re-try layout with exact margin-collapse result.

The process would be like below.
If there are dome structures,

<html>
  <body>
    <div>
    </div>
  </body>
</html>

The process to layout was changed from
FrameDocument layoutAll
  Framedocument layoutWidth(calculate margin)
    FrameBody layoutWidth(calculate margin)
            FrameBlockBox layoutWidth(calculate margin)
            FrameBlockBox marginCollapse
        FrameBody marginCollapse
    Framedocument layoutHeight(layoutBlock or layoutInline)
to
FrameDocument layoutAll
    Framedocument layoutWidth(calculate margin)
    Framedocument estimate position top
    Framedocument layoutHeight(layoutBlock or layoutInline)
        FrameBody layoutWidth(calculate margin)
        FrameBody estimate position top
        FrameBody layoutHeight(layoutBlock or layoutInline)
            FrameBlockBox layoutWidth(calculate margin)
            FrameBlockBox estimate position top
            FrameBlockBox layoutHeight(layoutBlock or layoutInline)
            FrameBlockBox relayout when triggered
...

Triggred factor to relayout is based on the assumption that between normal flowing blocks don't have to re-layout, but lastly there were
floating boxes, and the box to layout is higher than those of floating boxes, (which means y position of current box is less than
clearedDistanceToFloatBottom with BothClearValue) there has a possibility.

Besides that, if the block(A) clears floats and it is self-collapsing, using the expression below, the containing block's height
should be calculated.
Expression: max(max(max(margin-top, margin-bottom of next-self-collapsing blocks), margin-bottom of A) - margin-top of A, 0)