Fix bug of inserting pending inline box order
The order of pending inline boxes should be align with the order of inline boxes as they were pushed to line box.
Therefore, we should consider 2 cases, which there's nothing in pending inline boxes and there's some in pending inline boxes.
This commit considers both of cases, which in the past there weren't considered.
With a side effect, I also see the first white space in inline boxes due to re-layout can be removed at where we push the inline box
to the line box again.