Implement to jump to y position where float box should be located in
1. If there are only pending float box and the height which y diff for the next linebox is zero, then it goes infinite loop.
It happens because 2 things. First, the rule that float box can't be inserted into the line box when the last top position of float
box whose direction is the same with the direction of float box we are trying to insert. Second, this kind of float box can't be
detected by using current y position of line box and its height.
2. Fix minor bug of the target which getting the last top position of float box