Reimplement Float Layout
Before this commit, the layout of float is done after finishing inserting all the possible inline boxes and determining child inline
boxes' position. So after one line is finsihed to layout, then we finally register floating box to BlockFormattingContext.
But the problem of this implementation are 2 things.
The one is when small-height cleared floating boxes and some inline boxes share with one line box, we have to re-layout linebox so many times
trying to check if all the possible cleared floating boxes is fit to this line. We are not sure how many times, so we continuously tried
until the layout finally is settled. And this results in many redundant relayout because the floating boxes are not registered to
BlockFormattingContext so we have to repeat put and remove floating boxes and so on, in short word, it is not only no-intuitive, but also
difficult.
The other is we don't have to wait layout floating box until called breakLine or finishLine. Because no matter what happens, if floating box
is positioned, the position is never changed. Becuase floating box doesn't do vertical align or horizontal align.

With these 2 idea, I re-implemented the floating layout logic, if we can determine the position of floating in the early time, then we do.
Then the problem became very simple and we can also get benefits because we separate floating box from inline box. Many functions can be
simplified so we can focus on narrowed problem.

In the meantime, I refactored a little bit like below.
* Modifiy visibility of functions in LineFormattingContext.
* The x and width is changed to be calculated at only one function.
* Specify name with `reLayoutLineBoxIfNeeds` if it can relayout line box.
* Renaming if Singular or Plural form is more fit.