Fix layout bug of prefrerred width and computing width of percent value
CSS 2.1 spec says
" Note: For absolutely positioned elements whose containing block is based on a block container element, the percentage is calculated with
respect to the width of the padding box of that element. This is a change from CSS1, where the percentage width was always calculated with
respect to the content box of the parent element."

Therefore parent content width for floating box whose width is percent, it should be dealt with content box width.

Also, fix when computing preferred width and one of child boxes width is specified, don't loop over childs of that box.