Fix a bug related to Attribute List Selector
In case of [attr~=val] selector, it matches an element with the attr attribute
whose value is a whitespace-separated list of words, one of which is exactly
"val". If "val" contains whitespace, it will never represent anything (since
the words are separated by spaces). Also if "val" is the empty string, it will
never represent anything.