Modify logic of calculating heights for the IFC block
Originally the height of block which takes participation in inline formatting context was calculated collecting all the heights of line box.
So even though the line box is empty, once it contributes to calculate height, it cannot be removed. This commit changes this logic to
compute height of block measuring length from the bottom of the last line box or floating box to the top of the block. Then we don't need to
hold empty line box so that we just can remain only visible line boxes.