Remove code to find the valid last frame text
* `layoutInline` last white space shouldn't be layoutted, and this work was done at the very first time to start layoutInline, but it can
be replaced `removeDanglingSpace` when it is called at finishLine with flag `last line` on.
* Besides that, `InlineNonReplacedBox` has to be painted with starting and ending mbp. If `break` happened at last white space, the ending
mbp will be painted on the next line, which is a wrong result. So every time there has break line case due to white space, let it be
pending and wait until the next valid frame is met.

Signed-off-by: wonyong.kim wonyong.kim@samsung.com