Prepare to implement PrimaryGlobal extended attribute
Interface marked with `PrimaryGlobal` and attributes, or functions marked with `Unforgeable` should not be on the prototype chain. So if
any attributes or functions, it should be defined when generating object. And this is located in every
`virtual init(ScriptBindingInstance*)` function. So in order to generate automatically, I have to move codes to the each generated binding
code first. The real implementation for `PrimaryGlobal` will be done at the next commit.

Signed-off-by: wonyong.kim wonyong.kim@samsung.com