Generate attribute for Window
`Window` is the interface marked as `PrimaryGlobal`, so its attributes or functions should appear on the object itself not on the prototype
chain. So after `96597c5e`, moving `init` to inside the binding code, I manage to bind attributes or functions in `init` function. Also,
here I implemented `unforgeable`,some members of `Window` have, which means these members should not be forgeable, in other words, it is
not writable and not configurable.

is implmented in this commit, to

Signed-off-by: wonyong.kim wonyong.kim@samsung.com