Fix bug when navigating and needs to reset window
Everytime navigating, `window` should be reset with new `GlobalObject`. So I reset it calling `scriptObject()`. But even though
`scriptObject()` seems to reset object, but it doens't do actaully at the moment. Because I think it is not fully implemented to check if
it is valid or not. But sooner or later we have to implement to do that because like this case, the module can be different at different
`iframe` or navigation. Before landing this, I refreshed module myself forcely to use `scriptObjectSlowCase()`.
Also, `m_object` is handled different for `Window` on navigating, because it doesn't create new `Window` so that the assertion on the
`m_object` holding pointer to `ScriptWrappable` is broken. So for the `window`, we don't take `extraPointerData` instead take `this`
inside `scriptObjectSlowCase()`.

Signed-off-by: wonyong.kim wonyong.kim@samsung.com