Sync the index between utf16 and utf32 (#191)
icu library uses utf16 encoding to tokenize the text and internally for
the convenience, we used utf32 at this moment. And at some times,
because of surrogates(where code point ranges from 0xd800 to 0xdbff)
need 2 pairs of utf16 encoded code point. If it happens, the index
between icu and starfish should be synced.

Signed-off-by: wonyong.kim wonyong.kim@samsung.com