Fix bug of not applying overflow:hidden (#212)
Our clipping does at paint stage. and the state of canvas was preserved
during the paint function under certain layout parent. If the frame
establishes new StackingContext, then the state of canvas can be lost.
This commit if we reaches the frame which establishes StackingContext,
we save somewhere this information, and applies it under the paint
function with new StackingContext.

Signed-off-by: wonyong.kim wonyong.kim@samsung.com