Delay resolving Font and Length which contains font relative (#679)
Before introducing viewport relative length unit such as vw, vh, vmin and vmax,
font and length, we can assumed as a fixed value. With this assumption, we also
named these kind of unit like EmToBeFixed, ExToBeFixed or RemToBeFixed and we
changed this unit to fixed value. But when ,for example, `font-size: 1vw` comes,
it won't allow to change any more. But this case is very rare, so we have to
separate the case if any `font-size` has viewport relative value to take our
old assumption in general environment.

Signed-off-by: wonyong.kim wonyong.kim@samsung.com