Fix youtube/tv bugs
* Remove assert statement
There's a possibility that `layout` under `quick layout` process
For example,

<!DOCTYPE html>
<html lang="en">
<head>
</head>
<body onclick="f();">
    <div id="a" style="position:relative; width:200px; height: 200px;">
        <div id="b" style="float:left; width:100px; height:100px"></div>
        <div id="c" style="overflow: hidden; width: auto;">
          <div id="d" style="position:absolute; background: tomato">
            should layout
          </div>
        </div>
      </div>
</body>
<script>
  function f() {
    var span = document.querySelector('#b');
    b.style.width = '120px';
  };
</script>
</html>

in this case, if div whose id is 'c' is the candidate for quick layout
after f() is executed. Because from 'c''s perspective, at the steps of
`resolve style` and `build frame tree`, no one influenced, so it thinks
itself to skip layout. But from 'd''s perspective, as it is absoluted
box, `left` is auto, `right` is auto, and `width` is auto. Therefore
its content width depend on its x position to div whose id is 'a',
a.k.a, containing block. But no one can make sure its left position is
stable. So it should layout no matter what before it takes quick layout
step. (AFAIK, this is the only chance to do so.)

* Fix youtube/tv bug which causes invisible rendering,  when navigating
horizontally.

Signed-off-by: wonyong.kim wonyong.kim@samsung.com