Replace zeromq with nanomsg (#2421)
We've so far used zeromq for IPC. However, it's being used only for x86
linux mainly because of the license issue. For IPC between the main
process and workers, it's needed to replace it with the other solution.

This patch suggests nanomsg MIT-licensed. It's written by the original
author of zeromq. The following link describes the comparison with zmq
in detail: https://nanomsg.org/documentation-zeromq.html.
With nanomsg, Inspector can use web socket. The step to create the
nw-binding is consequently removed.

I will make another patch to clean up the zmq submodule once accepted.

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com