Add String::equal working with a given length (#2717)
Seeing the place where `String:: equal` is used through the codes,
a string literal is mostly given to the function. Since `string literal'
is `string constant` which is null-terminated, the current usage of
`String::equal` itself seems fine. Static analysis tool reports it may
occur a problem though.

However, it could be necessary to compare a data of String with variable
data. (e.g a data come from networking) This patch adds more functions
which receive additionally a length of the source string.

Signed-off-by: Daeyeon Jeong daeyeon.jeong@samsung.com