Fix memory alarm bugs (#3457)
* Fix memory alarm bugs

Whenever a linear gradient is drawn, the FrameBox continues to try to
cache the image data. However, if it fails due to a cache size limit,
the image data will accumulate until a GC is performed. This may cause
memory alarm problems. So, if the image data is not cached, remove it
immediately before doing GC.

Signed-off-by: Seungsoo Lee <seungsoo47.lee@samsung.com>

* Revise implementation to delete image data

Because of the shared_ptr, the destructor of a NativeGradient object is
called immediately when the object is deleted. So, I move
removeGradientImageDataCached() into the destructor instead of calling
it directly.

Signed-off-by: Seungsoo Lee <seungsoo47.lee@samsung.com>