Parse custom variables for transition property (#3516)
* Parse custom variables for transition property

The transition CSS property is a shorthand property for transition-property,
transition-duration, transition-timing-function, and transition-delay.
Each property (including longhand and shorthand properties) has to be
expressible using CSS custom variables.

Signed-off-by: Seungsoo Lee <seungsoo47.lee@samsung.com>

* Update transition() to handle custom variables

The function transition() does not deal with properly a case that transition
shorthand property includes custom variables. Also, the function is used to
return css value of computed style. But, to do this, style should be resolved first.

Signed-off-by: Seungsoo Lee <seungsoo47.lee@samsung.com>

* Remove unnecessary code

Signed-off-by: Seungsoo Lee <seungsoo47.lee@samsung.com>