Add X11 WebView bridge
The GLFW bridge simplifies window creation and GL settings by relying on
the framework, making it easy to start development. However, it lacks
transparency in controlling the details, which makes it difficult to
find out local environmental issues reported on WebGL.

A new port is introduced that directly combines X11 and EGL, removing
unclear aspects.

Signed-off-by: Daeyeon Jeong <daeyeon.jeong@samsung.com>