WebGL: Support EvasGL port (#3940)
This patch enables WebGL on the EFL/EvasGL port. 

Unlike EGL, EvasGL has no API to get GL environment information in `Starfish::WebView`, and its GL surface needs to be updated on the window resize event. To synchronize the information, this shares the `XGLPlatform` with `LWE::WebView` instead of creating a new mechanism. Additionally, this addresses pending refactorings from the previous PR.

Signed-off-by: Daeyeon Jeong <daeyeon.jeong@samsung.com>