Fix window creation for EFL port (#4148)
Goal:
The callbacks registered in `evas_object_image_pixels_get_callback_set`
are used for the gl rendering and composition. When we make the evas
image object dirty, the registered callback should be called.

Problem:
The problem is that for unknown reasons (timing issue, Elementary bug,
etc.), if the window isn't created properly, the `EVAS_CALLBACK_MOVE`
event isn't received, and the registered callbacks are not invoked.

Observation:
One thing I noticed is that when moving the position of the window
Evas_Object to the x:0, y:0 by default, if we set the window object
with a full-screen size (or any large size), then `EVAS_CALLBACK_MOVE`
event is not fired. In normal cases, the event should be received.

Fix:
This patch moves the window object slightly and then reverts it. This
is probably harmless because it's called before the window is shown.
This ensures that `EVAS_CALLBACK_MOVE` is received and pixels-get
callbacks are called correctly.

Signed-off-by: Daeyeon Jeong <daeyeon.jeong@samsung.com>